+++ src/Castle.Core/DynamicProxy/IInvocation.cs
@@ -115,6 +115,12 @@ namespace Castle.DynamicProxy
 		void Proceed();
 
 		/// <summary>
+		///   Returns an object describing the <see cref="Proceed"/> operation for this <see cref="IInvocation"/>
+		///   at this specific point during interception.
+		/// </summary>
+		IInvocationProceedInfo CaptureProceedInfo();
+
+		/// <summary>
 		///   Overrides the value of an argument at the given <paramref name = "index" /> with the
 		///   new <paramref name = "value" /> provided.
 		/// </summary>